langpack services module destroyed from mir_app the only survived service moved to mir_core
git-svn-id: http://svn.miranda-ng.org/main/trunk@14250 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/ICQCorp/src/protocol.cpp
@@ -2202,7 +2202,7 @@ void ICQ::addUrl(ICQUser *u, char *m, unsigned short theCmd, unsigned short theS
     CCSDATA ccs;
     PROTORECVEVENT pre;
 
-    messageLen = mir_tstrlen(m);
+    messageLen = (int)mir_tstrlen(m);
     for (i=0; i<messageLen; i++) if (m[i] == (char)0xFE) m[i] = 0;
 
     url = new char[messageLen + 1];
